Rebuilt the code's MD file to Jupyter Notebook
type: chore

The goal here was to put the original code into a more reproducible
format for other users. The Jupyter Notebook is easier to pick up
and work with and has great inline graphing features for dat
exploration.

In addition to flipping the code from a markdown file to a Jupyter
Notebook, there were some adjustments to the GridSearchCV. This
was mostly for users to see what the ideal parameters were for the
chosen model before being implemented on training and testing
sets