Visualization Addition
type: feat

The goal of this update is to implement visuals to the model producing
process to help users understand the different model transfers made.
Using myplotlib we are able to visualize each model produced and
compare, side by side how they stack up on the training and testing
datasets.