Including Gradient Descent Rrgressor
type: feat

Included in this update are two new regressor. The first is a classic
stochastic gradient descent algorithm that is tuned with GridSearchCV.
This will act as a baseline for the Boosted Gradient Regressor, which
we can expect to perform better (with MSE and r2 Scores as our
measure of goodness). Initial findings show that the sparsity of the
data is throwing off accuracy, so this may not be an ideal regressor
to continue working with.

There were also small inclusions in the README file to accomodate the
new regression methods.